pyproject.toml
[tool.poetry]
name = "uploading_processing_files"
version = "0.1.0"
description = ""
authors = ["Vera Dobrodeeva <dobro.vera@gmail.com>"]
packages = [
  { include = "uploading_processing" },
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
Markdown = "^3.4.4"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
celery = "^5.3.4"
redis = "^5.0.1"
importlib-metadata = "4.13.0"
coverage = "^7.3.1"
Pillow = "^10.0.1"
docker = "^6.1.3"
pipreqs = "^0.4.13"
docker-compose = "^1.29.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
